\name{exModelFile}
\alias{exModelFile}
\title{
Example model file
}
\usage{
exModelFile(modelName)
}
\arguments{
\item{modelName}{Model name (without '.lua' extension}
}
\description{
Gives full path to an example lua model file.
If no model name is given, the full list of available models is printed
out on screen.
}
\author{
Antonio, Fabio Di Narzo
}
\keyword{ misc }
\seealso{
}
\examples{
exModelFile()
exModelFile('logistic')
}

