\name{RiDMC-package}
\alias{RiDMC-package}
\alias{RiDMC}
\docType{package}
\title{
RiDMC: the iDMC->R interface
}
\description{
RiDMC: the iDMC->R interface
}
\details{
\code{RiDMC} is an R package released under GPL v.2 license for the 
numerical analysis of dynamical systems.

It uses the \code{idmclib} library.
}
\author{
Antonio, Fabio Di Narzo

Maintainer: Antonio, Fabio Di Narzo <antonio.fabio@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link{Model}} for a starting point.
}
