\name{internals}
\alias{RiDMC-internals}
\alias{buildModel}
\alias{checkModelDiscrete}
\alias{checkModelContinuous}
\alias{checkModelParVar}
\alias{checkPositiveScalar}
\alias{getModelType}
\alias{getModelName}
\alias{getModelDescription}
\alias{getModelParNames}
\alias{getModelVarNames}
\alias{getModelText}
\alias{getModelHasInverse}
\alias{getModelHasJacobian}
\alias{getBasinModel}
\alias{getBasinData}
\alias{getTrajectoryModel}
\alias{getTrajectoryValues}
\alias{idmclibVersion}
\alias{colorLegendGrob}
\alias{contentsGrob}
\alias{drawDetails.imageMap}
\alias{editDetails.colorLegendGrob}
\alias{editDetails.image}
\alias{editDetails.imageScale}
\alias{editDetails.plotGrob}
\alias{extend}
\alias{extend.grob}
\alias{as.grob}
\alias{as.grob.idmc_basin}
\alias{as.grob.idmc_bifurcation}
\alias{as.grob.idmc_lexp_diagram}
\alias{getField}
\alias{getField.grob}
\alias{Respect}
\alias{Respect.grob}
\alias{Respect<-}
\alias{Respect<-.grob}
\alias{Xlim}
\alias{Xlim.grob}
\alias{Xlim<-}
\alias{Xlim<-.grob}
\alias{Ylim}
\alias{Ylim.grob}
\alias{Ylim<-}
\alias{Ylim<-.grob}
\alias{grid.colorLegend}
\alias{grid.imageMap}
\alias{heightDetails.colorLegendGrob}
\alias{imageGrob}
\alias{imageMapGrob}
\alias{imageScaleGrob}
\alias{makeBasinsPalette}
\alias{makePlotGrobViewports}
\alias{mkPlotChildsAndViewports}
\alias{ribbonKids}
\alias{ribbonLegend}
\alias{ribbonVps}
\alias{tsGrob}
\alias{widthDetails.colorLegendGrob}
\alias{widthDetails.ribbonLegend}
\alias{xyGrob}
\alias{xyPlotGrob}
\alias{print.idmc_ctrajectory}
\alias{print.idmc_dtrajectory}
\alias{print.idmc_bifurcation}
\title{
RiDMC internal functions
}
\description{
RiDMC internal functions
}
\keyword{ internal }
