\name{internals}
\alias{RiDMC-internals}
\alias{checkModelParVar}
\alias{checkPositiveScalar}
\alias{getModelType}
\alias{getModelName}
\alias{getModelDescription}
\alias{getModelParNames}
\alias{getModelVarNames}
\alias{getModelText}
\alias{getModelHasInverse}
\alias{getModelHasJacobian}
\alias{getBasinModel}
\alias{getBasinData}
\alias{getTrajectoryModel}
\alias{getTrajectoryValues}
\alias{buildModel}
\alias{checkModelDiscrete}
\alias{checkModelContinuous}
\title{
RiDMC internal functions
}
\description{
RiDMC internal functions
}
\keyword{ internal }
