\name{RiDMC-package}
\alias{RiDMC-package}
\alias{RiDMC}
\docType{package}
\title{
RiDMC: the iDMC->R interface
}
\description{
RiDMC: the iDMC->R interface
}
\details{
\tabular{ll}{
Package: \tab RiDMC\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2007-05-19\cr
License: \tab GPL v.2\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Antonio, Fabio Di Narzo

Maintainer: Antonio, Fabio Di Narzo <antonio.fabio@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
